Race bar plot

Race bar plot

plot_race_wdi_ind(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  regions = wdiquickplots:::regions,
  income_groups = wdiquickplots:::income_groups,
  p = 1
)

Arguments

indicator
character of length 1 with the indicator code

highlight_countries
character vector with country names to highlight

start
first year to download data

end
last year to download data

country
subset of countries to download data from. Default to "all". This argument is passed as is to WDI::WDI and hence, requires 'ISO-2 character codes, e.g. "BR", "US", "CA"'.

regions
character vector to filter the data only to specific regions

income_groups
character vector to filter the data only to specific sincome groups

p
transformation exponent, as in scales::modulus_trans

Value

gganimate